Revert "Worked on random. Started converting the methodFrame stack and other things to unsigned types"
This reverts commit 21ef6b4fe238cccf85dd705947bfc87d94abe939.

+++ JavaVirtualMachine/HeapArray.cs
@@ -87,11 +87,11 @@ namespace JavaVirtualMachine
             }
             else if (Array is double[] doubleArr)
             {
-                doubleArr[index] = Utility.StoredDoubleToDouble((ulong)itemData);
+                doubleArr[index] = Utility.StoredDoubleToDouble(itemData);
             }
             else if (Array is float[] floatArr)
             {
-                floatArr[index] = (float)Utility.StoredDoubleToDouble((ulong)itemData);
+                floatArr[index] = (float)Utility.StoredDoubleToDouble(itemData);
             }
             else
             {
